Split middleware to submodules
