+++ go.work
@@ -1,0 +1,9 @@
+go 1.22.1
+
+use (
+	.
+	./middleware/ginz
+	./middleware/gorillaz
+	./middleware/grpcz
+	./middleware/httpz
+)
