+++ middleware/ginz/policy.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/aserto-dev/go-aserto/middleware"
-	httpmw "github.com/aserto-dev/go-aserto/middleware/http"
+	httpmw "github.com/aserto-dev/go-aserto/middleware/httpz"
 	"github.com/aserto-dev/go-aserto/middleware/internal"
 	authz "github.com/aserto-dev/go-authorizer/aserto/authorizer/v2"
 	"github.com/gin-gonic/gin"
