+++ middleware/grpcz/go.mod
@@ -1,0 +1,43 @@
+module github.com/aserto-dev/go-aserto/middleware/grpcz
+
+go 1.21
+
+replace github.com/aserto-dev/go-aserto => ../../
+
+require (
+	github.com/aserto-dev/go-aserto v0.32.0
+	github.com/aserto-dev/go-authorizer v0.20.8
+	github.com/aserto-dev/go-directory v0.31.9
+	github.com/lestrrat-go/jwx/v2 v2.1.1
+	github.com/pkg/errors v0.9.1
+	github.com/samber/lo v1.46.0
+	github.com/stretchr/testify v1.9.0
+	google.golang.org/grpc v1.65.0
+	google.golang.org/protobuf v1.34.2
+)
+
+require (
+	buf.build/gen/go/bufbuild/protovalidate/protocolbuffers/go v1.34.2-20240717164558-a6c49f84cc0f.2 // indirect
+	github.com/aserto-dev/errors v0.0.9 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.3.0 // indirect
+	github.com/goccy/go-json v0.10.3 // indirect
+	github.com/grpc-ecosystem/grpc-gateway/v2 v2.20.0 // indirect
+	github.com/lestrrat-go/blackmagic v1.0.2 // indirect
+	github.com/lestrrat-go/httpcc v1.0.1 // indirect
+	github.com/lestrrat-go/httprc v1.0.6 // indirect
+	github.com/lestrrat-go/iter v1.0.2 // indirect
+	github.com/lestrrat-go/option v1.0.1 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rs/zerolog v1.33.0 // indirect
+	github.com/segmentio/asm v1.2.0 // indirect
+	golang.org/x/crypto v0.25.0 // indirect
+	golang.org/x/net v0.27.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	golang.org/x/text v0.16.0 // indirect
+	google.golang.org/genproto/googleapis/api v0.0.0-20240722135656-d784300faade // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240722135656-d784300faade // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
