+++ middleware/httpz/export_test.go
@@ -1,4 +1,4 @@
-package http
+package httpz
 
 func InternalHostnameSegment(hostname string, level int) string {
 	return hostnameSegment(hostname, level)
