+++ middleware/httpz/go.mod
@@ -1,0 +1,28 @@
+module github.com/aserto-dev/go-aserto/middleware/httpz
+
+go 1.21
+
+replace github.com/aserto-dev/go-aserto => ../../
+
+require (
+	github.com/aserto-dev/go-aserto v0.32.0
+	github.com/aserto-dev/go-authorizer v0.20.8
+	github.com/lestrrat-go/jwx/v2 v2.1.1
+	gotest.tools v2.2.0+incompatible
+)
+
+require (
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.3.0 // indirect
+	github.com/goccy/go-json v0.10.3 // indirect
+	github.com/google/go-cmp v0.6.0 // indirect
+	github.com/lestrrat-go/blackmagic v1.0.2 // indirect
+	github.com/lestrrat-go/httpcc v1.0.1 // indirect
+	github.com/lestrrat-go/httprc v1.0.6 // indirect
+	github.com/lestrrat-go/iter v1.0.2 // indirect
+	github.com/lestrrat-go/option v1.0.1 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/segmentio/asm v1.2.0 // indirect
+	golang.org/x/crypto v0.25.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
+)
