+++ middleware/httpz/identity.go
@@ -1,4 +1,4 @@
-package http
+package httpz
 
 import (
 	"net/http"
