+++ middleware/internal/contextutil.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-func ValueOrEmpty(ctx context.Context, key interface{}) string {
+func ValueOrEmpty(ctx context.Context, key any) string {
 	identity, ok := ctx.Value(key).(string)
 	if ok {
 		return identity
